Python code for JetRacer closes #8
