--- nano/camera/face-detect-usb.py
@@ -1,74 +1,0 @@
-
-
-import cv2
-
-window_title = "Face Detect"
-
-def face_detect():
-    face_cascade = cv2.CascadeClassifier(
-        "/usr/share/opencv4/haarcascades/haarcascade_frontalface_default.xml"
-    )
-    eye_cascade = cv2.CascadeClassifier(
-        "/usr/share/opencv4/haarcascades/haarcascade_eye.xml"
-    )
-    # ASSIGN CAMERA ADDRESS HERE
-    camera_id="/dev/video0"
-    # Full list of Video Capture APIs (video backends): https://docs.opencv.org/3.4/d4/d15/group__videoio__flags__base.html
-    # For webcams, we use V4L2
-    video_capture = cv2.VideoCapture(camera_id, cv2.CAP_V4L2)
-    """ 
-    # How to set video capture properties using V4L2:
-    # Full list of Video Capture Properties for OpenCV: https://docs.opencv.org/3.4/d4/d15/group__videoio__flags__base.html
-    #Select Pixel Format:
-    # video_capture.set(cv2.CAP_PROP_FOURCC, cv2.VideoWriter_fourcc(*'YUYV'))
-    # Two common formats, MJPG and H264
-    # video_capture.set(cv2.CAP_PROP_FOURCC, cv2.VideoWriter_fourcc(*'MJPG'))
-    # Default libopencv on the Jetson is not linked against libx264, so H.264 is not available
-    # video_capture.set(cv2.CAP_PROP_FOURCC, cv2.VideoWriter_fourcc(*'H264'))
-    # Select frame size, FPS:
-    video_capture.set(cv2.CAP_PROP_FRAME_WIDTH, 640)
-    video_capture.set(cv2.CAP_PROP_FRAME_HEIGHT, 480)
-    video_capture.set(cv2.CAP_PROP_FPS, 30)
-    """
-
-    if video_capture.isOpened():
-        try:
-            cv2.namedWindow(window_title, cv2.WINDOW_AUTOSIZE)
-            # Start counting the number of frames read and displayed
-            # left_camera.start_counting_fps()
-            while True:
-                _, frame = video_capture.read()
-                gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
-                faces = face_cascade.detectMultiScale(gray, 1.3, 5)
-
-                for (x, y, w, h) in faces:
-                    cv2.rectangle(frame, (x, y), (x + w, y + h), (255, 0, 0), 2)
-                    roi_gray = gray[y : y + h, x : x + w]
-                    roi_color = frame[y : y + h, x : x + w]
-                    eyes = eye_cascade.detectMultiScale(roi_gray)
-                    for (ex, ey, ew, eh) in eyes:
-                        cv2.rectangle(
-                            roi_color, (ex, ey), (ex + ew, ey + eh), (0, 255, 0), 2
-                        )
-
-                # Check to see if the user closed the window
-                # Under GTK+, WND_PROP_VISIBLE does not work correctly. Under Qt it does
-                # GTK - Substitute WND_PROP_AUTOSIZE to detect if window has been closed by user
-                if cv2.getWindowProperty(window_title, cv2.WND_PROP_AUTOSIZE) >= 0:
-                    cv2.imshow(window_title, frame)
-                else:
-                    break
-                
-                keyCode = cv2.waitKey(30) & 0xFF
-                # Stop the program on the ESC key or 'q'
-                if keyCode == 27 or keyCode == ord('q'):
-                    break
-        finally:
-            video_capture.release()
-            cv2.destroyAllWindows()
-    else:
-        print("Unable to open camera")
-
-
-if __name__ == "__main__":
-    face_detect()
