--- nano/camera/teste.py
@@ -1,18 +1,0 @@
-import cv2
-
-cap = cv2.VideoCapture(0)
-cap.set(cv2.CAP_PROP_CONVERT_RGB, 1)
-cap.set(cv2.CAP_PROP_FRAME_WIDTH, 1280)
-cap.set(cv2.CAP_PROP_FRAME_HEIGHT, 720)
-
-while True:
-    ret, frame = cap.read()
-    if ret:
-        frame = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
-        cv2.imshow('Camera', frame)
-    
-    if cv2.waitKey(1) & 0xFF == ord('q'):
-        break
-
-cap.release()
-cv2.destroyAllWindows()
